Combine two utils files related to knex
