+++ api/lib/infrastructure/repositories/schooling-registration-repository.js
@@ -8,7 +8,7 @@ const studentRepository = require('./student-repository');
 const Bookshelf = require('../bookshelf');
 const BookshelfSchoolingRegistration = require('../data/schooling-registration');
 const bookshelfToDomainConverter = require('../utils/bookshelf-to-domain-converter');
-const bookshelfUtils = require('../utils/bookshelf-utils');
+const bookshelfUtils = require('../utils/knex-utils');
 
 function _toUserWithSchoolingRegistrationDTO(BookshelfSchoolingRegistration) {
